+++ lib/grpc.dart
@@ -13,6 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+export 'package:grpc/src/generated/google/protobuf/any.pb.dart';
 export 'package:grpc/src/generated/google/rpc/error_details.pb.dart';
 
 export 'src/auth/auth.dart' show BaseAuthenticator;
@@ -42,7 +43,8 @@ export 'src/client/proxy.dart' show Proxy;
 export 'src/client/transport/http2_credentials.dart'
     show BadCertificateHandler, allowBadCertificates, ChannelCredentials;
 export 'src/server/call.dart' show ServiceCall;
-export 'src/server/interceptor.dart' show Interceptor;
+export 'src/server/interceptor.dart'
+    show Interceptor, ServerInterceptor, ServerStreamingInvoker;
 export 'src/server/server.dart'
     show
         ServerCredentials,